Add GameSessionScreen for unified game sessions
