+++ navigation/AppStack.js
@@ -7,13 +7,12 @@ import ChatScreen from '../screens/ChatScreen';
 import NotificationsScreen from '../screens/NotificationsScreen';
 import EditProfileScreen from '../screens/EditProfileScreen';
 import GameInviteScreen from '../screens/GameInviteScreen';
-import GameLobbyScreen from '../screens/GameLobbyScreen';
+import GameSessionScreen from '../screens/GameSessionScreen';
 import CommunityScreen from '../screens/CommunityScreen';
 import EventChatScreen from '../screens/EventChatScreen';
 import PremiumScreen from '../screens/PremiumScreen';
 import PremiumPaywallScreen from '../screens/PremiumPaywallScreen';
 import StatsScreen from '../screens/StatsScreen';
-import GameWithBotScreen from '../screens/GameWithBotScreen';
 
 const Stack = createNativeStackNavigator();
 
@@ -26,13 +25,13 @@ export default function AppStack() {
       <Stack.Screen name="Notifications" component={NotificationsScreen} />
       <Stack.Screen name="EditProfile" component={EditProfileScreen} />
       <Stack.Screen name="GameInvite" component={GameInviteScreen} />
-      <Stack.Screen name="GameLobby" component={GameLobbyScreen} />
+      <Stack.Screen name="GameLobby" component={GameSessionScreen} />
       <Stack.Screen name="Community" component={CommunityScreen} />
       <Stack.Screen name="EventChat" component={EventChatScreen} />
       <Stack.Screen name="Premium" component={PremiumScreen} />
       <Stack.Screen name="PremiumPaywall" component={PremiumPaywallScreen} />
       <Stack.Screen name="Stats" component={StatsScreen} />
-      <Stack.Screen name="GameWithBot" component={GameWithBotScreen} />
+      <Stack.Screen name="GameWithBot" component={GameSessionScreen} />
     </Stack.Navigator>
   );
 }
